Move WorkFilter into its own file
